+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 polls Changelog
 ===============
 
+Version 1.1.0 stable (2013-11-26)
+---------------------------------
+
+### Improved
+- Added the option to reset poll (thanks to René Zschoch)
+- Added the multiple vote feature (thanks to René Zschoch)
+
+
 Version 1.0.2 stable (2013-09-16)
 ---------------------------------